+++ src/components/cookie/cookie.tsx
@@ -1,7 +1,7 @@
 import {
   Component, h, Prop, State, Element, Watch
 } from '@stencil/core';
-import { themeStyle } from '../../helpers/themeStyle.js';
+import { themeStyle } from '../../helpers/themeStyle';
 
 import JsCookie from 'js-cookie';
 import Tab from 'bootstrap/js/src/tab';
