+++ src/components/footer/footer.tsx
@@ -1,7 +1,7 @@
 import {
   Component, h, Prop, State, Element, Watch,
 } from '@stencil/core';
-import { themeStyle } from '../../helpers/themeStyle.js';
+import { themeStyle } from '../../helpers/themeStyle';
 
 @Component({
   tag: 'c-footer',
