+++ src/components/icon/icon.tsx
@@ -1,7 +1,7 @@
 import {
   Component, h, Prop, State, Watch, Element,
 } from '@stencil/core';
-import { themeStyle } from '../../helpers/themeStyle.js';
+import { themeStyle } from '../../helpers/themeStyle';
 
 @Component({
   tag: 'c-icon',
